import csv files done

+++ routes/web.php
@@ -42,8 +42,9 @@ Route::resource('/student/project', 'StudentProjectsController')->middleware('au
 // Admin
 //Route::group(['middleware' => ['auth','admin']], function (){
     Route::get('/admin', 'AdminController@index', ['name' => 'tables']);
-    Route::get("/edit-user/{id}",'AdminController@edit');
-    Route::put("/edit-user-update/{id}",'AdminController@update');
+    Route::get("/admin/edit-user/{id}",'AdminController@edit');
+    Route::put("/admin/edit-user-update/{id}",'AdminController@update');
     Route::resource('/admin', 'AdminController');
-    Route::delete('/delete-user/{id}','AdminController@destroy');
+    Route::delete('/admin/delete-user/{id}','AdminController@destroy');
+    Route::post('/admin/import/','AdminController@import');
 //});
